backend/requirements.txt
########################################################################
#  Alya Ultimate – Backend requirements (18 avril 2025)
########################################################################
# 1 · FastAPI & Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
orjson==3.10.16
ujson==5.10.0
httpx==0.27.0

# 2 · LLM  (OpenAI + Azure)
openai==1.25.1
azure-identity==1.16.0
azure-core==1.30.1
azure-ai-ml==1.14.0                # Agent Service SDK
azure-search-documents==11.6.0b1   # Vector Search / Mémoire
tiktoken==0.6.0

# 3 · Cloud SDKs & Storage
azure-storage-blob==12.19.1
azure-keyvault-secrets==4.8.0
msal==1.32.0                       # Auth Entra ID / Graph
msgraph-core==1.3.3                # Microsoft Graph SDK (GA)

# 4 · Intégrations SAAS
hubspot-api-client==9.0.0
py-trello==0.19.0
slack_sdk==3.27.2
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# 5 · Worker & Data
celery==5.4.0
redis==5.0.4
psycopg[binary]==3.1.18
sqlalchemy==2.0.29
alembic==1.13.1
sqlalchemy==2.0.29

# 6 · File‑handling / Parsing
python-magic==0.4.27
python-jose[cryptography]==3.3.0
alembic==1.13.1
pandas==2.2.1
openpyxl==3.1.2
python-docx==1.0.1
pypdf2==3.0.1
xlrd==2.0.1
beautifulsoup4==4.12.3

# 7 · Utils & Observability
loguru==0.7.2
tqdm==4.67.1
python-multipart==0.0.20
email-validator==2.2.0
prometheus-client==0.20.0
sentry-sdk==2.4.0
ratelimit==2.2.1
########################################################################
#  Fin – prévois ~3‑5 min pour `docker compose build backend`
########################################################################

